Episode 24 Complete

+++ app/Http/Controllers/ThreadController.php
@@ -109,6 +109,7 @@ class ThreadController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize("update", $thread);
         $thread->delete();
 
         return redirect('/threads');
